requirements.txt
altgraph==0.17
Desktopmagic==14.3.11
future==0.18.2
pefile==2019.4.18
Pillow==7.0.0
PyQt5==5.14.1
PyQt5-sip==12.7.2
pytesseract==0.3.3
pywin32==227
pywin32-ctypes==0.2.0
